# Refactor MiqDialog.seed for performance improvements