--- a/spec/models/miq_dialog/data/product/dialogs/miq_dialogs/testing.yaml
+++ b/spec/models/miq_dialog/data/product/dialogs/miq_dialogs/testing.yaml
@@ -1,0 +1,23 @@
+---
+:name: testing_dialog
+:description: Testing Dialog
+:dialog_type: MiqProvisionWorkflow
+:content:
+  :dialog_order:
+  - :requester
+  - :service
+  :buttons:
+  - :continue
+  - :cancel
+  :dialogs:
+    :requester:
+      :description: Request
+      :fields: {}
+
+      :display: :hide
+      :field_order:
+    :service:
+      :description: Catalog
+      :fields: {}
+
+      :display: :hide