… spec/models/miq_dialog_spec.rb