Initial seed changed from 0.69s to 0.72s (-4.3% savings - insignificant)
Second seed dropped from 1.40s to 0.014s (99% savings)

This PR is similar to #18335 in just about every way.

@jrafanie @NickLaMuro Please review.

@gmcculloug Please also review as this makes the changes to destroy removed dialogs as we had discussed.