@jrafanie 2019-01-22
@Fryguy are you going to change any rubocops or @NickLaMuro's comment re: YAML load being the part we want to avoid?  If not, I'll merge.  It looks like everything else was addressed.