@Fryguy 2019-01-23
@jrafanie I'm going to leave the existing things...this is good to merge...I'm considering opening some guides votes for the things above.